requirements.txt
kfp==2.0.1
tensorflow==2.12.0
pandas
scikit-learn
joblib
google-cloud-aiplatform
